Extract and reuse ControlledSearchInput component
