+++ apps/charterafrica/src/components/Datasets/DatasetFilterBar.js
@@ -9,7 +9,7 @@ import {
 import React, { useState } from "react";
 
 import { neutral } from "@/charterafrica/colors";
-import SearchInput from "@/charterafrica/components/SearchInput";
+import { ControlledSearchInput } from "@/charterafrica/components/SearchInput";
 
 const StyledAutocomplete = styled(Autocomplete)({
   backgroundColor: neutral[50],
@@ -40,46 +40,33 @@ const DatasetFilterBar = React.forwardRef(function DatasetFilterBar(
   ref
 ) {
   const {
-    onQChange,
-    onSortChange,
+    onChangeQ,
+    onChangeSort,
     onChangeCountries,
     onChangeTags,
-    search: searchProp,
     countries,
     labels,
-    tags,
     sortOptions,
+    q,
+    tags,
   } = props;
 
-  const [value, setValue] = useState(searchProp || "");
   const [selectedCountries, setSelectedCountries] = useState([
     labels.countries,
   ]);
   const [selectedTags, setSelectedTags] = useState([labels.tags]);
 
-  const handleChangeQ = (e) => {
-    setValue(e.target.value);
-  };
-
-  const handleClick = (e) => {
-    if (onQChange) {
-      onQChange(e, value);
-    }
-  };
-
-  const handleKeyPress = (e) => {
-    if (e.key === "Enter") {
-      if (onQChange) {
-        onQChange(e, value);
-      }
+  const handleChangeQ = (e, value) => {
+    if (onChangeQ) {
+      onChangeQ(e, value);
     }
   };
 
   const handleChangeSort = (e, selectedSortOption) => {
     if (selectedSortOption?.value) {
-      onSortChange(e, selectedSortOption.value);
+      onChangeSort(e, selectedSortOption.value);
     } else {
-      onSortChange(e, "");
+      onChangeSort(e, "");
     }
   };
 
@@ -129,12 +116,10 @@ const DatasetFilterBar = React.forwardRef(function DatasetFilterBar(
     <Box py={5} ref={ref}>
       <Grid container spacing={2}>
         <Grid item xs={12} lg={3}>
-          <SearchInput
+          <ControlledSearchInput
             onChange={handleChangeQ}
-            onClick={handleClick}
-            onKeyPress={handleKeyPress}
             placeholder={labels.search}
-            value={value}
+            value={q}
             sx={{
               backgroundColor: "#fff",
               height: "36px",
