+++ apps/charterafrica/src/components/SearchInput/index.js
@@ -1,3 +1,5 @@
 import SearchInput from "./SearchInput";
 
+export { default as ControlledSearchInput } from "./ControlledSearchInput";
+
 export default SearchInput;
